#162684426 Change code file structure to improve readability and improve modularity
