+++ app/api/v1/__init__.py
@@ -1,20 +1,14 @@
 from flask_restful import Api, Resource
 from flask import Blueprint
 
-from .views import Interventions
-from .views import Redflags
-from .views import Intervention
-from .views import Redflag
-from .views import Signup
-from .views import Login
-from .views import UpdateInterventionLocation
-from .views import UpdateRedflagLocation
-from .views import UpdateInterventionComment
-from .views import UpdateRedflagComment
-from .views import Interventionstatus
-from .views import Redflagstatus
-from .views import Protected
-
+from .incidents.redflag_views import Redflag, Redflags, Redflagstatus, \
+                                      UpdateRedflagComment, \
+                                      UpdateRedflagLocation
+from .incidents.intervention_views import Intervention, Interventions, \
+                                          Interventionstatus, \
+                                          UpdateInterventionComment, \
+                                          UpdateInterventionLocation
+from .users.users_views import Signup, Login
 
 version_one = Blueprint('api_v1', __name__, url_prefix='/api/v1')
 api = Api(version_one)
@@ -36,4 +30,3 @@ api.add_resource(UpdateRedflagComment,
 api.add_resource(Interventionstatus,
                  '/interventions/<intervention_id>/status')
 api.add_resource(Redflagstatus, '/redflags/<redflag_id>/status')
-api.add_resource(Protected, '/protected')
